Make IP checking optional

+++ app/Http/Controllers/Auth/AbstractLoginController.php
@@ -86,14 +86,16 @@ abstract class AbstractLoginController extends Controller
 
         $this->clearLoginAttempts($request);
 
-        $ip = User::where('ip', $request->getClientIp())->count();
+        if ($this->settings->get('jexactyl::registration:ip_check') == 'true') {
+            $ip = User::where('ip', $request->getClientIp())->count();
 
-        if ($ip > 1) {
-            try {
-                // Attempt to delete the user when alting detected.
-                $user->delete();
-            } catch (Exception $ex) { /* do nothing */ }
-            throw new DisplayException('You have been detected using an alt account. Your IP address ('.$request->getClientIp().') has been logged.');
+            if ($ip > 1) {
+                try {
+                    // Attempt to delete the user when alting detected.
+                    $user->delete();
+                } catch (Exception $ex) { /* do nothing */ }
+                throw new DisplayException('You have been detected using an alt account. Your IP address ('.$request->getClientIp().') has been logged.');
+            }
         }
 
         $this->auth->guard()->login($user, true);
